reuse code with UserDictBuilder
