+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "lindera",
     "lindera-analyzer",
+    "lindera-dictionary-builder",
     "lindera-cc-cedict",
     "lindera-cc-cedict-builder",
     "lindera-cli",
@@ -25,6 +26,7 @@ resolver = "2"
 [workspace.dependencies]
 lindera = { version = "0.30.0", path = "lindera" }
 lindera-analyzer = { version = "0.30.0", path = "lindera-analyzer" }
+lindera-dictionary-builder = { version = "0.30.0", path = "lindera-dictionary-builder" }
 lindera-cc-cedict = { version = "0.30.0", path = "lindera-cc-cedict" }
 lindera-cc-cedict-builder = { version = "0.30.0", path = "lindera-cc-cedict-builder" }
 lindera-cli = { version = "0.30.0", path = "lindera-cli" }
@@ -60,9 +62,9 @@ log = "0.4.21"
 once_cell = "1.19.0"
 rand = "0.8.5"
 regex = "1.10.3"
-serde = {version="1.0.197", features = ["derive"] }
+serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.114"
-strum = {version="0.26.2", features = ["derive"] }
+strum = { version = "0.26.2", features = ["derive"] }
 strum_macros = "0.26.2"
 tar = "0.4.40"
 thiserror = "1.0.58"
@@ -71,6 +73,7 @@ unicode-normalization = "0.1.23"
 unicode-segmentation = "1.11.0"
 ureq = { version = "2.9.6", default-features = false, features = ["tls"] }
 yada = "0.5.1"
+derive_builder = "0.20.0"
 
 
 [profile.release]
