+++ lindera-cc-cedict-builder/Cargo.toml
@@ -28,3 +28,4 @@ yada.workspace = true
 lindera-compress = { workspace = true, optional = true }
 lindera-core.workspace = true
 lindera-decompress.workspace = true
+lindera-dictionary-builder.workspace = true
